Change query modules structures
