+++ opac/queries/__init__.py
@@ -3,5 +3,4 @@ from opac.queries.errors import *  # noqa: F401 F403
 from opac.queries.holding import *  # noqa: F401 F403
 from opac.queries.lending import *  # noqa: F401 F403
 from opac.queries.renewing import *  # noqa: F401 F403
-from opac.queries.reservation import *  # noqa: F401 F403
 from opac.queries.stock import *  # noqa: F401 F403
