+++ opac/queries/holding/__init__.py
@@ -1,0 +1,2 @@
+from .lend import *  # noqa: F401 F403
+from .cancel import *  # noqa: F401 F403
