+++ opac/queries/lending/__init__.py
@@ -1,0 +1,1 @@
+from .back import *  # noqa: F401 F403
