+++ opac/queries/stock/__init__.py
@@ -1,0 +1,1 @@
+from .first_reservation_to_holding import *  # noqa: F401 F403
